backend/requirements.txt
asgiref==3.5.2
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.1.2
django-colorfield==0.7.2
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.1
fonttools==4.38.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.2
Pillow==9.2.0
pycparser==2.21
pydyf==0.5.0
PyJWT==2.5.0
pyphen==0.13.0
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.5
requests==2.28.1
requests-oauthlib==1.3.1
simplejwt==2.0.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tinycss2==1.2.1
typing==3.7.4.3
tzdata==2022.5
uritemplate==4.1.1
urllib3==1.26.12
weasyprint==57.1
webencodings==0.5.1
zopfli==0.2.1
